requirements.txt
python-socketio==1.6.1
eventlet==0.20.0
gevent==1.2.0
gevent-websocket==0.9.5
Flask==0.12
gunicorn==19.6.0
